+++ packages/client/src/network/explorer/goals.ts
@@ -1,0 +1,11 @@
+import { World } from '@mud-classic/recs';
+
+import { Components } from 'network/';
+import { getAllGoals, getGoalByIndex } from 'network/shapes/Goal';
+
+export const goals = (world: World, components: Components) => {
+  return {
+    all: () => getAllGoals(world, components),
+    get: (index: number) => getGoalByIndex(world, components, index),
+  };
+};
