+++ packages/client/src/network/shapes/Config/index.ts
@@ -1,4 +1,9 @@
 export { getConfig as getKamiConfig } from './kami';
-export { getConfigFieldValue, getConfigFieldValueArray, getConfigFieldValueWei } from './types';
+export {
+  getConfigFieldValue,
+  getConfigFieldValueAddress,
+  getConfigFieldValueArray,
+  getConfigFieldValueWei,
+} from './types';
 
 export type { Config as KamiConfig } from './kami';
